+++ sglang_omni/serve/openai_api.py
@@ -91,21 +91,13 @@ def create_app(
     _register_chat_completions(app)
     _register_speech(app)
     try:
-        from sglang_omni.serve.webrtc_api import create_realtime_router
-
-        app.include_router(
-            create_realtime_router(client, model_name=app.state.model_name)
-        )
-    except Exception:
-        logger.exception("Failed to register realtime prototype routes")
-    try:
         from sglang_omni.serve.realtime_ws_api import create_realtime_ws_router
 
         app.include_router(
             create_realtime_ws_router(client, model_name=app.state.model_name)
         )
     except Exception:
-        logger.exception("Failed to register realtime websocket routes")
+        logger.exception("Failed to register realtime routes")
 
     return app
